+++ lib/home_screen.dart
@@ -253,7 +253,7 @@ class _HomeScreenState extends State<HomeScreen> {
                                 Align(
                                   alignment: Alignment.centerLeft,
                                   child: Text(
-                                    "App version: Beta 04.03",
+                                    "App version: Beta 04.04",
                                     style: TextStyle(
                                       color: darkenColor(
                                         appColorNotifier.value,